+++ test/models/user.rb
@@ -9,6 +9,7 @@ class User
   validates_confirmation_of :password
 
   validates_length_of :name, minimum: 10, maximum: 100
+  validates_length_of :password, is: 8
   validates_size_of :lastname, maximum: 100
 
   validates_presence_of :name