+++ CHANGELOG.md
@@ -13,3 +13,4 @@
 + Add `with_max` and `is_at_most` aliases to `ValidateLengthMatcher#with_maximum`.
 + Add `within` option to `ValidateLengthMatcher`.
 + Add `in` alias to `ValidateLengthMatcher#within`.
++ Add `is` option to `ValidateLengthMatcher`.